feat(docker): Create feature flag for Docker
This refactors the Docker configuration by creating a new `docker` feature flag in `features/options.nix`.

This feature enables the Docker service and adds the primary user to the `docker` group, centralizing the logic and removing duplication.

The `gamey-wsl`, `thinky-wsl`, and `teeny` systems have been updated to use this new feature.